requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.5
h11==0.14.0
httptools==0.6.4
idna==3.10
openpyxl==3.1.5
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pymongo==4.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1
pandas~=2.2.3
flask~=3.1.0